Improve text (#1)
* Unify spelling to "E-Mail"

Now "E-Mail" is "E-Mail", and not "Email" or "email".
Note that "Schul-E-Mail" looks a bit ridiculous, but it's more correct than "Schulemail". (Which can be read as "Schule|mail" which is weird too.)

* Improve text slightly

There were a few typos and grammatical errors. Also, I added some punctuation.

* Match capitalization of example mail to real mail addresses

Most mail addresses are capitalized on the Birklehof addressbook on Outlook.